przerobienie całego kodu, żeby działał na Androidzie

+++ src/main/java/technology/tabula/Cell.java
@@ -1,6 +1,5 @@
 package technology.tabula;
 
-import java.awt.geom.Point2D;
 import java.util.Collections;
 
 @SuppressWarnings("serial")
@@ -13,7 +12,9 @@ public class Cell extends RectangularTextContainer<TextChunk> {
 	}
 
 	public Cell(Point2D topLeft, Point2D bottomRight) {
-		super((float) topLeft.getY(), (float) topLeft.getX(), (float) (bottomRight.getX() - topLeft.getX()), (float) (bottomRight.getY() - topLeft.getY()));
+		super(topLeft.getY(), topLeft.getX(),
+				bottomRight.getX() - topLeft.getX(),
+				bottomRight.getY() - topLeft.getY());
 		this.setPlaceholder(false);
 		this.setSpanning(false);
 	}
